Dynamic SPLADE batch size

+++ langchain_websearch.py
@@ -5,6 +5,7 @@ import logging
 
 import aiohttp
 import requests
+import torch
 from bs4 import BeautifulSoup
 from langchain.retrievers.document_compressors import DocumentCompressorPipeline
 from langchain.retrievers.ensemble import EnsembleRetriever
@@ -47,11 +48,11 @@ class LangchainCompressor:
             self.splade_doc_tokenizer = AutoTokenizer.from_pretrained("naver/efficient-splade-VI-BT-large-doc",
                                                                       cache_dir=model_cache_dir)
             self.splade_doc_model = AutoModelForMaskedLM.from_pretrained("naver/efficient-splade-VI-BT-large-doc",
-                                                                         cache_dir=model_cache_dir).to(self.device)
+                                                                         cache_dir=model_cache_dir, torch_dtype=torch.float16).to(self.device)
             self.splade_query_tokenizer = AutoTokenizer.from_pretrained("naver/efficient-splade-VI-BT-large-query",
                                                                         cache_dir=model_cache_dir)
             self.splade_query_model = AutoModelForMaskedLM.from_pretrained("naver/efficient-splade-VI-BT-large-query",
-                                                                           cache_dir=model_cache_dir).to(self.device)
+                                                                           cache_dir=model_cache_dir, torch_dtype=torch.float16).to(self.device)
             optimum_logger = optimum.bettertransformer.transformation.logger
             original_log_level = optimum_logger.level
             # Set the level to 'ERROR' to ignore "The BetterTransformer padding during training warning"
